+++ src/main/java/vectorwing/blockbox/common/tag/CompatibilityTags.java
@@ -13,6 +13,10 @@ public class CompatibilityTags
 	public static final String FARMERS_DELIGHT = "farmersdelight";
 	public static final TagKey<Block> FARMERS_DELIGHT_TRAY_HEAT_SOURCES = externalBlockTag(FARMERS_DELIGHT, "tray_heat_sources");
 
+	public static final String SABLE = "sable";
+	public static final TagKey<Block> SABLE_SUPER_LIGHT = externalBlockTag(SABLE, "super_light");
+	public static final TagKey<Block> SABLE_LIGHT = externalBlockTag(SABLE, "light");
+
 	private static TagKey<Block> externalBlockTag(String modId, String path) {
 		return BlockTags.create(ResourceLocation.fromNamespaceAndPath(modId, path));
 	}